error404-backend/requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn
anyio
openai
jiter
pydantic-core